Only watch for entries in packages/ directory, prevent recursion
If a package contains a lot of files (for example @mdi/svg under
node_modules/), the number of watchers that are created is very large.

Depending on the system configuration, the following error may get
reported, which causes CamScripter to error out:

    node:internal/errors:464
        ErrorCaptureStackTrace(err);
        ^

    Error: ENOSPC: System limit for number of file watchers reached, watch '/usr/local/lib/node_modules/camscripter-raspberry/packages/CamStreamDeckControl/node_modules/pako/lib/zlib/gzheader.js'
        at FSWatcher.<computed> (node:internal/fs/watchers:244:19)

There is no need to watch the contents of the micro app, reducing the
depth to 1 (instead of the default 'undefined') only allocates 8
watchers, where with the default value over 12000 are created.